Update to AddTheme, Made GetActiveThemeID
Made an update to Addtheme Now it also makes an Themes_Calander Line with the Time_Slot in which the theme is active. The GetActiveThemeID Gets the theme that has an timeslot in which the datetime.Today falls, and then it will return the id of that theme if none are found then it will return null.